Element base and derived refactored -nNodes added

+++ src/domain/Topology.h
@@ -9,9 +9,9 @@
 
 namespace Topology{
 
-template<unsigned int Dim> static constexpr bool EmbeddableInSpace{Dim<4};
-template<unsigned int Dim> static constexpr bool InPlane = Dim==2; // Unused Yet 
-template<unsigned int Dim> static constexpr bool InSpace = Dim==3; // Unused Yet
+template<unsigned short Dim> static constexpr bool EmbeddableInSpace{Dim<4};
+template<unsigned short Dim> static constexpr bool InPlane = Dim==2; // Unused Yet 
+template<unsigned short Dim> static constexpr bool InSpace = Dim==3; // Unused Yet
 
 
 template<unsigned int Dim1, unsigned int Dim2=Dim1>
